ai-service/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
google-generativeai==0.7.2